HS-106-Summarize

+++ src/main/java/edu/hm/cs/seng/hypershop/service/UnitConversionService.java
@@ -17,7 +17,7 @@ public class UnitConversionService {
         customUnits = new HypershopCustomUnits();
     }
 
-    IngredientAmount summarizeIngredients(final IngredientAmount ingredient, int amount, String unitName) {
+    IngredientAmount summarizeIngredients(final IngredientAmount ingredient, double amount, String unitName) {
 
         if (canSummarize(ingredient.getUnit(), unitName)) {
             Unit baseUnit = getUnit(ingredient.getUnit());
